Prevent admin from deleting roles that are already assigned to users

+++ server/controllers/role.controller.js
@@ -1,5 +1,6 @@
 import mongoose from "mongoose";
 import Role from "../models/role.model.js";
+import User from "../models/user.model.js";
 import * as permissionCheck from "../util/permissionCheck.js";
 import { canPerformAction } from "../util/utils.js";
 
@@ -52,6 +53,12 @@ export const deleteRole = async (req, res) => {
 
         if (!mongoose.Types.ObjectId.isValid(roleId)) {
             return res.status(404).json({ error: "No roles found with that id" });
+        }
+
+        const totalUserWithThisRole = await User.find({ roleId }).count();
+
+        if (totalUserWithThisRole > 0) {
+            return res.status(405).json({ message: `Forbidden: ${totalUserWithThisRole} user(s) has this role. ` });
         }
 
         await Role.deleteOne({ _id: roleId });
